Move PlaybookBuilder class to ssg module
